refactor: move pipeline/learning.rs into astra-turn-core
Move learning.rs (1,184 LOC) to astra-turn-core::pipeline_learning.
This module bridges TurnLearningWriter trait to pipeline modules.

The turn↔pipeline dependency cycle is broken because:
- astra-turn-core already depends on astra-pipeline
- learning_quality_gate was moved to turn-core (PR #128)
- turn types (contracts, stall) are in turn-core

Changes:
- Add tracing dependency to astra-turn-core
- Move learning.rs with fixed imports
- Add #[allow(deprecated)] to test module
- Replace runtime stub with re-export

Depends on: PR #128 (learning_quality_gate extraction)
