+++ rust/crates/astra-turn-core/Cargo.toml
@@ -27,6 +27,7 @@ serde_json = { workspace = true }
 serde_yaml = "0.9.34"
 tokio = { workspace = true }
 tokio-util = { workspace = true }
+tracing = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
 
 [dev-dependencies]
